geospatial/geopandas/requirements.txt
geopandas==0.14.0
pandas==2.1.3
pandas-gbq==0.19.2
google-cloud-bigquery==3.13.0
shapely==2.0.2
pyproj==3.6.1
fiona==1.9.5
